Add stick navbar and desktop styles
