+++ src/components/events/eventsPage.js
@@ -4,6 +4,6 @@ function EventsPage() {
       <h1>Events Page</h1>
     </div>
   );
-};
+}
 
 export default EventsPage;
