+++ src/components/footer/footerSection.js
@@ -1,0 +1,9 @@
+function FooterSection() {
+  return (
+    <div className="footer-section">
+      <h1>Footer Section</h1>
+    </div>
+  );
+}
+
+export default FooterSection;
