+++ src/components/organizers/organizersPage.js
@@ -4,6 +4,6 @@ function OrganizersPage() {
       <h1>Organizers Page</h1>
     </div>
   );
-};
+}
 
 export default OrganizersPage;
